fix: Provide proper implementation if planned process monitoring
Initial functionality was to simply use the terminate
functionality with the liveview which proved not
to be correct way to do it. Here is a proper process
monitoring setup that will check if the liveview
process is still running and update the engagement
time accordingly.

With the fix to properly monitor the process we can
remove the need for hacky way of doing the redirect
and simplify the liveview modules.